ai/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
python-dotenv>=0.19.0
mysql-connector-python>=8.0.0
python-dateutil>=2.8.2
SQLAlchemy>=1.4.0
flask==2.3.3
requests==2.31.0
